Move geo annotation to the annotation directory
