+++ frontend/src/components/annotation/geo_annotation/geo_hotkeys.vue
