+++ frontend/src/components/annotation/geo_annotation/geo_sidebar.vue
@@ -148,11 +148,11 @@
 
 <script lang="ts">
 import Vue from 'vue'
-import instance_detail_list_view from "../annotation/image_annotation/instance_detail_list_view.vue";
-import button_with_menu from '../regular/button_with_menu.vue';
-import label_select_only from '../label/label_select_only.vue'
-import attribute_group_list from '../attribute/attribute_group_list.vue'
-import global_attributes_list from '../attribute/global_attributes_list.vue'
+import instance_detail_list_view from "../../annotation/image_annotation/instance_detail_list_view.vue";
+import button_with_menu from '../../regular/button_with_menu.vue';
+import label_select_only from '../../label/label_select_only.vue'
+import attribute_group_list from '../../attribute/attribute_group_list.vue'
+import global_attributes_list from '../../attribute/global_attributes_list.vue'
 
 export default Vue.extend({
     name: "geo_sidepanel",
