+++ frontend/src/components/annotation/geo_annotation/geo_tile.vue
